requirements.txt
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-groq
pypdf
faiss-cpu
streamlit
openai
sentence-transformers
python-dotenv
